CR feedback re: moving actions.ts functions to usage components where possible
