+++ packages/components/src/entities/index.ts
@@ -1,3 +1,5 @@
+import { PICKLIST_SAMPLES_FILTER } from './models';
+import { getSampleTypes, getOriginalParentsFromLineage } from './actions';
 import {
     getSampleWizardURL,
     filterSampleRowsForOperation,
@@ -5,7 +7,6 @@ import {
     isFindByIdsSchema,
     getSampleDeleteMessage,
     getSampleTypeTemplateUrl,
-    getOriginalParentsFromLineage,
     createEntityParentKey,
 } from './utils';
 import { SampleTypeBasePage } from './SampleTypeBasePage';
@@ -43,9 +44,7 @@ import { PicklistListing } from './PicklistListing';
 import { PicklistOverview } from './PicklistOverview';
 import { PicklistSubNav } from './PicklistSubnav';
 import { SamplesTabbedGridPanel } from './SamplesTabbedGridPanel';
-import { SampleTypeTemplateDownloadRenderer } from './SampleTypeTemplateDownloadRenderer';
-import { PICKLIST_SAMPLES_FILTER } from './models';
-import { downloadSampleTypeTemplate, getSampleTypes } from './actions';
+import { SampleTypeTemplateDownloadRenderer, downloadSampleTypeTemplate } from './SampleTypeTemplateDownloadRenderer';
 
 export {
     PICKLIST_SAMPLES_FILTER,
